+++ src/ResourceManager/Insights/Commands.Insights.Test/Alerts/AddAzureRmLogAlertRuleTests.cs
@@ -18,6 +18,8 @@ using Microsoft.Azure.Commands.ResourceManager.Common;
 using Microsoft.Azure.Commands.ScenarioTest;
 using Microsoft.Azure.Management.Insights;
 using Microsoft.Azure.Management.Insights.Models;
+using Microsoft.Azure.Management.Monitor.Management;
+using Microsoft.Azure.Management.Monitor.Management.Models;
 using Microsoft.WindowsAzure.Commands.ScenarioTest;
 using Moq;
 using System;
@@ -34,7 +36,7 @@ namespace Microsoft.Azure.Commands.Insights.Test.Alerts
     public class AddAzureRmLogAlertRuleTests
     {
         private readonly AddAzureRmLogAlertRuleCommand cmdlet;
-        private readonly Mock<InsightsManagementClient> insightsManagementClientMock;
+        private readonly Mock<MonitorManagementClient> insightsManagementClientMock;
         private readonly Mock<IAlertRulesOperations> insightsAlertRuleOperationsMock;
         private Mock<ICommandRuntime> commandRuntimeMock;
         private Rest.Azure.AzureOperationResponse<AlertRuleResource> response;
@@ -46,15 +48,21 @@ namespace Microsoft.Azure.Commands.Insights.Test.Alerts
             TestExecutionHelpers.SetUpSessionAndProfile();
             // XunitTracingInterceptor.AddToContext(new XunitTracingInterceptor(output));
             insightsAlertRuleOperationsMock = new Mock<IAlertRulesOperations>();
-            insightsManagementClientMock = new Mock<InsightsManagementClient>();
+            insightsManagementClientMock = new Mock<MonitorManagementClient>();
             commandRuntimeMock = new Mock<ICommandRuntime>();
             cmdlet = new AddAzureRmLogAlertRuleCommand()
             {
                 CommandRuntime = commandRuntimeMock.Object,
-                InsightsManagementClient = insightsManagementClientMock.Object
+                MonitorManagementClient = insightsManagementClientMock.Object
             };
 
-            AlertRuleResource alertRuleResourceInput = new AlertRuleResource(location: null, isEnabled: true, alertRuleResourceName: "a name");
+            AlertRuleResource alertRuleResourceInput = new AlertRuleResource()
+            {
+                Location = null, 
+                IsEnabled = true, 
+                AlertRuleResourceName = "a name"
+            };
+
             response = new Rest.Azure.AzureOperationResponse<AlertRuleResource>()
             {
                 Body = alertRuleResourceInput