+++ src/ResourceManager/Insights/Commands.Insights.Test/Autoscale/NewAzureRmAutoscaleProfileTests.cs
@@ -17,6 +17,7 @@ using Microsoft.Azure.Commands.Insights.Autoscale;
 using Microsoft.Azure.Commands.ResourceManager.Common;
 using Microsoft.Azure.Commands.ScenarioTest;
 using Microsoft.Azure.Management.Insights.Models;
+using Microsoft.Azure.Management.Monitor.Management.Models;
 using Microsoft.WindowsAzure.Commands.ScenarioTest;
 using Moq;
 using System;