+++ src/ResourceManager/MachineLearning/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Consume new version of Azure Machine Learning .Net SDK and add a new cmdlet
+    - Add-AzureRmMlWebServiceRegionalProperty 
+* Minor wording fixes in help text.
 
 ## Version 0.13.0